Module sda_protocol::helpers

Various helper traits, structs, methods, and macros.

Structs

Binary
Blob of binary data.

Labelled
Abstract object for messages labelled by some form of identifier.

Signed
Abstract object for signed message together with the signature and claimed signer.

Traits

Id
Trait for unique identifiers.

Identified
Trait for objects with associated unique identifiers.

Sign
Abstract trait for objects that may be signed.

Functions

label