Struct sda_protocol::Aggregation

pub struct Aggregation {
    pub id: AggregationId,
    pub title: String,
    pub vector_dimension: usize,
    pub modulus: i64,
    pub recipient: AgentId,
    pub recipient_key: EncryptionKeyId,
    pub masking_scheme: LinearMaskingScheme,
    pub committee_sharing_scheme: LinearSecretSharingScheme,
    pub recipient_encryption_scheme: AdditiveEncryptionScheme,
    pub committee_encryption_scheme: AdditiveEncryptionScheme,
}

Description of an aggregation.

Fields
id: AggregationId

Unique identifier of aggregation.

title: String

Title, e.g. explaining purpose of the aggregation.

vector_dimension: usize

Fixed dimension of input and output vectors.

modulus: i64

The group in which all operations are performed.

Cryptographic primitives must be compatible with this value.

recipient: AgentId

Recipient of output vector.

recipient_key: EncryptionKeyId

Encryption key to be used for encryptions intended for the recipient.

masking_scheme: LinearMaskingScheme

Scheme and parameters used for masking secrets between the recipient and the committee.

committee_sharing_scheme: LinearSecretSharingScheme

Scheme and parameters used for sharing masked secrets between the clerks in the committee.

recipient_encryption_scheme: AdditiveEncryptionScheme

Scheme and parameters used for encrypting masks for the recipient.

committee_encryption_scheme: AdditiveEncryptionScheme

Scheme and parameters used for encrypting shares of masked secrets for the committee.